+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - `XML`
   - `PLIST`
 - Add support for specifying persistence method as index in addition to name.
+- Add new `-n/--nuke` option to delete payload and Nekrosis on exit.
 
 ## 0.0.1
 - Initial release.